refactor: consolidate formatDuration to single source in time-format.ts
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)
