+++ src/features/background-agent/result-handler.ts
@@ -1,5 +1,5 @@
 export type { ResultHandlerContext } from "./result-handler-context"
-export { formatDuration } from "./duration-formatter"
+export { formatDuration } from "../../tools/background-task/time-format"
 export { getMessageDir } from "./message-storage-locator"
 export { checkSessionTodos } from "./session-todo-checker"
 export { validateSessionHasOutput } from "./session-output-validator"
