+++ src/tools/delegate-task/sync-continuation.ts
@@ -7,7 +7,7 @@ import { getAgentToolRestrictions } from "../../shared/agent-tool-restrictions"
 import { getMessageDir } from "../../shared/session-utils"
 import { promptWithModelSuggestionRetry } from "../../shared/model-suggestion-retry"
 import { findNearestMessageWithFields } from "../../features/hook-message-injector"
-import { formatDuration } from "./time-formatter"
+import { formatDuration } from "../background-task/time-format"
 import { syncContinuationDeps, type SyncContinuationDeps } from "./sync-continuation-deps"
 import { setSessionTools } from "../../shared/session-tools-store"
